fix big-endian conversion for uint64, add support for floating point
when converting to network order on a little endian machine, the upper and lower
bits of a 64-bit integer where switched the wrong way around. this commit fixes
that and adds support for doubles and floats. the testcase is pretty redundant
though.

floating point support makes assumption that float is 4 byte, double is 8
byte. refer to corresponding static assert.

because testing private methods is a bit controversial, this has been omitted,
hopefully other failures if any will be found regardless.